% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bwNRD0}
\alias{bwNRD0}
\title{Silverman's Rule of Thumb Bandwidth for Kernel Density Estimation}
\usage{
bwNRD0(y, m)
}
\arguments{
\item{y}{a data vector}

\item{m}{a multiplier to adjust the SROT proportionally.}
}
\description{
A C++ version of Silverman's rule-of-thumb bandwidth. This is similar
with R's \code{bw.nrd0(x)}
}
\examples{
data(lba)
h <-cpda::bwNRD0(plba$DT1, 0.8)

}

