% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pLBA.R
\name{Compute_log_likelihood_FFT}
\alias{Compute_log_likelihood_FFT}
\title{Compute Log-likelihood Using Fast Fourier Transform}
\usage{
Compute_log_likelihood_FFT(Subject_data, params, MCMC_params)
}
\arguments{
\item{Subject_data}{subject data}

\item{params}{LBA parameters}

\item{MCMC_params}{MCMC parameters}
}
\value{
Log-likelihood
}
\description{
This function takes subject data, the parameters for the current chain
under consideration, and MCMC parameters and computes the log likelihood.
}

