% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpda.R
\docType{package}
\name{cpda}
\alias{cpda}
\alias{cpda-package}
\title{Probability Density Approximation}
\description{
This package implements probability density approximation in C++
}
\references{
Holmes, W. (2015). A practical guide to the Probability Density
Approximation (PDA) with improved implementation and error characterization.
\emph{Journal of Mathematical Psychology}, \bold{68-69}, 13--24,
doi: http://dx.doi.org/10.1016/j.jmp.2015.08.006.
}
\author{
Yi-Shin Lin <yishin.lin@utas.edu.au> \cr
Andrew Heathcote <andrew.heathcote@utas.edu.au> \cr
William Holmes <william.holmes@vanderbilt.edu>
}
\keyword{cpda}
