% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cquantile}
\alias{cquantile}
\title{A simple and fast quantile calculator}
\usage{
cquantile(y, q)
}
\arguments{
\item{y}{a data vector}

\item{q}{nth quantile. Enter proportion, such as .25 or .75.}
}
\description{
A C++ version quantile function.
}

