% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossover}
\alias{crossover}
\title{DEMC Algorithm}
\usage{
crossover(useTheta, gamma, k, rp)
}
\arguments{
\item{useTheta}{a npar x nchain matrix}

\item{gamma}{a gamma vector.}

\item{k}{current chain index. Note chain index starts from 0.}

\item{rp}{ter Braak's (2006) b, setting the range of the uniform 
distribution that derived epsilon. This is usually set 0.001 or suggested 
by ter Braak (ROT) 1e-4.}
}
\value{
a column vector
}
\description{
This is part of DEMC algorithm. This function uses crossover sampler to 
proposal a new parameter vector.
}
\examples{
pVec <- c(A=1.51, b=2.7, muv1=3.32, muv2=2.24, t_ND=0.08, muw1=1.51,  
          muw2=3.69, t_delay=0.31, sv=1, swt=0.5)
setting  <- c(bandwidth=.02, ns=1e4, nmc=6, nchain=24, rp=.001, burnin=10,
              nthin=3, start=1, gammaMult=2.38, report=20)
samples  <- init(pVec, setting)
useTheta <- samples$theta[,,1]
dim(useTheta)  ## 10 x 24; npar x nchain

gamma <- gammaVec(length(pVec), setting[9])

## Current chain index is 0 (C-based index)
crossover(useTheta, gamma, 0, 1e-4)

}
\keyword{a}
\keyword{matrix}

