% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lba-data.R
\docType{data}
\name{d}
\alias{d}
\title{Probability Density Approximation Data Sets}
\source{
\href{http://dx.doi.org/10.1016/j.jmp.2015.08.006}{PDA paper}
}
\description{
An example data set to fit piecewise LBA model, converted from Holmes's 
(2015) MATLAB formatted Example 4 Files (Data1.mat).
}
\examples{
\dontrun{
data(lba)
head(d)
}
##   Response ResponseTime Block
## 1        0    0.5676343     1
## 2        1    0.6183177     1
## 3        0    0.8806298     1
}
\keyword{dataset}

