% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pLBA.R
\name{DE_MCMC_2block_fun}
\alias{DE_MCMC_2block_fun}
\title{Calculate Gaussian Log-likelihood and Probability Density}
\usage{
DE_MCMC_2block_fun(MCMC_params, Subject_data, Model_specifics, savename,
  report = 100)
}
\arguments{
\item{MCMC_params}{Provides all MCMC and KDE parameters along with a
few other things}

\item{Subject_data}{Stores the RT data to be fit.}

\item{Model_specifics}{Stores function handles and block information
for the model}

\item{savename}{Specifies where you want the MCMC output stored.}

\item{report}{the iteration interval of returning a progress report.
Default 100}

\item{...}{other arguments}
}
\value{
the chain parameter information (param_chain). This ends up
saved in a data file however. The matrix has a structure
\code{param_chain(MCMC_iteration,parameter num,chain num)}
}
\description{
This function takes all relevant inputs from the driver and runs a
2-block, differential evolution MCMC
}
\references{
Holmes, W. (2015). A practical guide to the Probability Density
Approximation (PDA) with improved implementation and error characterization.
\emph{Journal of Mathematical Psychology}, \bold{68-69}, 13--24,
doi: http://dx.doi.org/10.1016/j.jmp.2015.08.006.
}
\keyword{DE_MCMC_2block_fun}

