% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dinvG}
\alias{dinvG}
\alias{pinvG}
\alias{qinvG}
\alias{rinvG}
\title{The inverse-Gaussian Distribution}
\usage{
dinvG(x, mu = 1, lambda = 1, log = FALSE)

pinvG(q, mu = 1, lambda = 1, lower.tail = TRUE, log.p = FALSE)

qinvG(p, mu = 1, lambda = 1, lower.tail = TRUE, log.p = FALSE)

rinvG(n, mu = 1, lambda = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{vector of mu parameters.}

\item{lambda}{vector of lambda parameters.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as \code{log(p)}}

\item{lower.tail}{logical; if TRUE, (default), probabilities are 
\eqn{P[X \le x]} otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dinvG} gives the density. \code{pinvG} gives the distribtuion
function, \code{qinvG} gives the quantile function, and \code{rinvG} 
generates random deviates. \eqn{mu <= 0} or \eqn{lambda <= 0} is an error.
}
\description{
Density, distribution function, quantile function and random generation for 
the inverse-Gaussian distribution with \eqn{mean = mu} and 
\eqn{standard deviation} \eqn{ = mu^{3} / lambda}. See Table 1 in 
Cousineau, Brown, & Heathcote (2004) for the equation. kappa assumes 0.
}
\details{
The inverse-Gaussian is also known as the Wald distribution.
}
\examples{
x <- seq(0, 1, length.out = 1024)
den <- cpda::dinvG(x, mu=.275, lambda=2)
plot(x, den, type = "l", lwd = 2)

## Use Table 3 in Cousineau, Brown, & Heathcote (2004) as an example
n <- 4096
sam <- cpda::rinvG(n, 275, 2000) + 725
hist(sam, freq = FALSE, breaks = "FD", main = "Wald Distribution", 
  xlab = "RT (ms)", ylab="Density")
  
}
