% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gammavec}
\alias{gammavec}
\title{Generate a Gamma Vector}
\usage{
gammavec(npar, gammamult = 2.38)
}
\arguments{
\item{npar}{number of parameters.}

\item{gammamult}{a tuning parameter for gamma mutation. Default value is 2.38.}
}
\value{
a column vector
}
\description{
This is part of DEMC algorithm. \code{gammavec} generates a gamma vector 
for element-wise compuation in Armadillo C++.
}
\examples{
pVec <- c(A=1.51, b=2.7, muv1=3.32, muv2=2.24, t_ND=0.08, muw1=1.51,  
          muw2=3.69, t_delay=0.31,  sv=1, swt=0.5)
gamma <- gammavec(length(pVec), 2.38)
}
