% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pLBA.R
\name{get_bandwidth}
\alias{get_bandwidth}
\title{Compute Kernel Bandwidth}
\usage{
get_bandwidth(object, ns, ...)
}
\arguments{
\item{object}{a data vector s}

\item{ns}{number of simulation}

\item{...}{other arguments}
}
\value{
a scalar value of bandwidth
}
\description{
Use Silverman's run of thumb to calculate kernel bandwidth.
}
\examples{
datavec <- rnorm(1000, 5, 1)
get_bandwidth(datavec, 10000)
}

