% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getDTs}
\alias{getDTs}
\title{Retrieve Empirical and Modelled Decision Times}
\usage{
getDTs(data, pVec, MCMC_params)
}
\arguments{
\item{data}{a Rcpp List for subject data}

\item{pVec}{a Rcpp NumvericVector for parameter vector}

\item{MCMC_params}{MCMC parameters}
}
\value{
A four-element list with estimated decision time for accumulator 1,
accumultor 2, empirical decision time for accumualtor 1 and accumulator 2.
}
\description{
Get decision times for two-accumulator LBA model
}
\examples{
load("data/Data1.rda")
p.vector <- c(1.51, 3.32, 1.51, 2.24, 3.69, 0.31, 0.08)
mcmcParams <- list(sigma_exact=1, bandwidth=.02, LL_NSAMPLE=1e2,
                    Nstep=30, Nchain=24, noise_size=.001, burnin=10,
                    resample_mod=3)

DTs <- getDTs(d, p.vector, mcmcParams)
str(DTs)
}

