% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEdges}
\alias{getEdges}
\title{Calculate edges for simulated histogram}
\usage{
getEdges(z)
}
\arguments{
\item{z}{a grid a scalar, usually created by 
\code{z = arma::linspace<arma::vec>(z0, z1, 1<<(int)p);}}
}
\description{
This is an internal function. The user may not use it.
}
