% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPRT}
\alias{getPRT}
\title{Retrieve piecewise RTs from a pLBA Model}
\usage{
getPRT(data, pVec, MCMC_params)
}
\arguments{
\item{data}{a Rcpp List for subject data}

\item{pVec}{a Rcpp NumvericVector for parameter vector}

\item{MCMC_params}{MCMC parameters}
}
\value{
A two-element list
}
\description{
Get piecewise RTs
}
\examples{
load("data/Data1.rda")
p.vector <- c(1.51, 3.32, 1.51, 2.24, 3.69, 0.31, 0.08)
mcmcParams <- list(sigma_exact=1, bandwidth=.02, LL_NSAMPLE=1e2,
                    Nstep=30, Nchain=24, noise_size=.001, burnin=10,
                    resample_mod=3)

tmp0 <- getPRT(d, p.vector, mcmcParams)
str(tmp0[[1]])
str(tmp0[[2]])
}

