% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init}
\alias{init}
\title{Initialise a pLBA Bayeisan Sample}
\usage{
init(pVec, setting)
}
\arguments{
\item{pVec}{a vector storing pLBA model parameters. The sequence is 
critical. It is, A, b, muv1, muv2, t_ND, muw1, muw2, t_delay, sv, swt.}

\item{setting}{DE and MC setting, including, nmc, nchain, etc.}
}
\value{
a list with 7 elements: param_old, param_chain, proposal, direction,
LL_keep, nmc, and nchain
}
\description{
This functions initialises a pLBA sample
}
\examples{
pVec <- c(A=1.51, b=2.7, muv1=3.32, muv2=2.24, t_ND=0.08, muw1=1.51,  
          muw2=3.69, t_delay=0.31,  sv=1, swt=0.5)

## Setting sequence is critical, too! 
setting <- c(bandwidth=.02, ns=1e5, nmc=30, nchain=24, rp=.001, burnin=10,
nthin=3, start=1, gammaMult=2.38, report=100)
tmp0 <- init(pVec, setting)
str(tmp0)
## List of 5
## $ opVec    : num [1:10, 1:24] -1.749 3.207 2.432 4.913 0.391 ...
## $ theta    : num [1:10, 1:24, 1:30] -1.749 3.207 2.432 4.913 0.391 ...
## $ prospoal : num [1:10, 1:24] NA NA NA NA NA NA NA NA NA NA ...
## $ direction: num [1:10, 1:24] NA NA NA NA NA NA NA NA NA NA ...
## $ LL_keep  : num [1:30, 1:24] NA NA NA NA NA NA NA NA NA NA ...

}
\keyword{initialize_structures}

