% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pLBA.R
\name{initialize_structures}
\alias{initialize_structures}
\title{Initialize a Structure}
\usage{
initialize_structures(nmc = 100, npar = 7, nchain = 3, b = 2.7)
}
\arguments{
\item{nmc}{number of MCMC iteration (steps). DMC's nmc. Default 100}

\item{npar}{number of parameter. Default 2}

\item{nchain}{number of chains. Default 3}

\item{b}{LBA's b parameter. Default 2.7}
}
\value{
a list with 5 elements: param_old, param_chain, proposal, direction,
and LL_keep
}
\description{
This file initializes the initial conditions and various data structures
for the MCMC
}
\examples{
nchain <- 3
nmc    <- 100
npar   <- 7
initialize_structures(nmc, npar, nchain)
}
\keyword{initialize_structures}

