% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializeStructures}
\alias{initializeStructures}
\title{Initialize a DMC Sample}
\usage{
initializeStructures(nmc, npar, nchain)
}
\arguments{
\item{nmc}{number of MCMC iteration (steps).}

\item{npar}{number of parameter.}

\item{nchain}{number of chains}
}
\value{
a list with 7 elements: param_old, param_chain, proposal, direction,
LL_keep, nmc, and nchain
}
\description{
This functions initializes the initial conditions and various data structures
for the MCMC, using C++
}
\examples{
initializeStructures(nmc=20, npar=7, nchain=3)
}
\keyword{initialize_structures}

