% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.pda}
\alias{logLik.pda}
\title{Calculate Gaussian Log-likelihood and Probability Density}
\usage{
\method{logLik}{pda}(object, pVec, params, ...)
}
\arguments{
\item{object}{a data vector to be fit. I've sorted it internally.}

\item{pVec}{a parameter vector for the approximated likelihood}

\item{params}{a list carrying kernel density estimation parameters.
Nsample, bandwidth, and precision (power of 2)}

\item{...}{other arguments}
}
\value{
a list containing four elements: LL = Log likelihood, PDF = PDF of
the data values, grid_centers  = a histogram grid, and PDF_hist = PDF on
that histogram grid.
}
\description{
A \code{logLik} function to compute density and log-likelihood, using
probability density approximation.
}
\examples{
rm(list=ls())

## Simulate a data vector from a Gaussian distribution
params <- c(mu=5, sigma=1)
datavec <- rnorm(1000, params[1], params[2])

## 1. We begin by first drawing N_s = 10,000 from the underlying distribution
## to generate a synthetic data set X^~
ns <- 10000; ## Number of synthetic data samples used for likelihood estimation.
iqr <- IQR(datavec, type=1);
S <- sd(datavec);

mcmcParams <- list(Nsample=ns, ## Silvermans rule of thumb
                   bandwidth=0.9*min(iqr, S)*ns^(-0.2),
                   precision=10);

class(datavec) <- "pda"
ll <- logLik(datavec, params, mcmcParams);

end1 <- length(ll$PDF_hist)-1
## png(file="figs/pda.png", width=800,height=600)
plot(ll$gridCenters, ll$PDF_hist[1:end1], type="l", lty=2,
     main="Normal Distribution",xlab="x",ylab="L(x|beta)")
lines(datavec,ll$PDF, col="red", lwd = 1.5)
## dev.off()
str(ll)
}
\references{
Holmes, W. (2015). A practical guide to the Probability Density
Approximation (PDA) with improved implementation and error characterization.
\emph{Journal of Mathematical Psychology}, \bold{68-69}, 13--24,
doi: http://dx.doi.org/10.1016/j.jmp.2015.08.006.
}
\keyword{logLik.pLBA}

