% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_fft}
\alias{logLik_fft}
\alias{logLik_fft2}
\title{Compute Log-likelihood Using KDE-based Fast Fourier Transform}
\usage{
logLik_fft(y, yhat)

logLik_fft2(y, yhat)
}
\arguments{
\item{y}{a vector storing empirical data (e.g., RTs)}

\item{yhat}{a vector storing simulated data (e.g., simualted RTs, using a
LBA model).}
}
\value{
Log-likelihood
}
\description{
This function implements Holmes's (2015) KDE-FFT method to calculate
approximated probability density. The precision is set at 10 (ie 2^10).
\code{logLik_fft2} returns three additional vectors, PDF (discrete
probability densities for interpolated empirical data), grid centers, and
PDF_hist (discrete probability densities for the simulated data).
}
\examples{
## Use piecewise LBA data as an example
data(lba)
logLik_fft(plba$DT1, plba$eDT1)
logLik_fft(plba$DT2, plba$eDT2)
tmp1 <- logLik_fft2(plba$DT1, plba$eDT1)
tmp2 <- logLik_fft2(plba$DT2, plba$eDT2)

str(tmp1)
## List of 4
## $ LL      : num 278
## $ PDF     : num [1:695, 1] 2.441 0.632 1.966 2.359 2.214 ...
## $ z       : num [1:1024, 1] 0.0889 0.0903 0.0916 0.093 0.0943 ...
## $ PDF_hist: num [1:1024, 1] 0 0 0 0 0 0 0 0 0 0 ...
str(tmp2)
## List of 4
## $ LL      : num 42.9
## $ PDF     : num [1:305, 1] 1.879 0.965 1.834 0.326 0.921 ...
## $ z       : num [1:1024, 1] 0.067 0.0696 0.0722 0.0748 0.0774 ...
## $ PDF_hist: num [1:1024, 1] 0 0 0 0 0 0 0 0 0 0 ...
}
\references{
Holmes, W. (2015). A practical guide to the Probability Density
Approximation (PDA) with improved implementation and error characterization.
\emph{Journal of Mathematical Psychology}, \bold{68-69}, 13--24,
doi: http://dx.doi.org/10.1016/j.jmp.2015.08.006.
}

