% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_fft2}
\alias{logLik_fft2}
\title{Compute Summed Log-Likelihood}
\usage{
logLik_fft2(y, yhat, h = 0, m = 0.8, p = 10)
}
\arguments{
\item{y}{a scalar or vector storing empirical data.}

\item{yhat}{a scalar or vector storing simulated data (e.g., 
\code{rnorm(100)}).}

\item{h}{the bandwidth for kernel density estimation. If not given, 
\code{logLik_fft} will use Sliverman's rule of thumb}

\item{m}{a multiplier to adjust \code{h} proportationally. Default is 0.8. 
This applies also when the user enters his/her own bandwidth.}

\item{p}{a precision parameter defines the number of grid as power of 2.
Default value is 10 (i.e., 2^10).}
}
\value{
A list with 4 elements.
}
\description{
\code{logLik_fft2} uses an identical algorithm as \code{logLik_fft}, but
return more elements. Differing from \code{logLik_pw}, \code{logLik_fft2} 
and \code{logLie_fft},  
\enumerate{
\item takes Monte Carlo simulations,
\item transforms them to spectral domain using FFT,
\item applies a standard Gaussian kernel to smooth them,
\item transforms them back to signal domain and
\item interpolates linearly the simulation histogram to the observed data 
points to obtain estimated likelihoods.
}
 
\code{logLik_fft2} returns four elements:
\itemize{
\item \bold{\emph{LL}}, summed, logged likelihood. This is identical as the 
return value from \code{logLik_fft}. 
\item \bold{\emph{PDF}}, a matrix storing individual data point (column 1)
and logged probability densities (column 2).
\item \bold{\emph{z}}, a vector storing centre points of the 
simulated histogram (i.e., grid centre)
\item \bold{\emph{PDF_hist}} a vector stoing the count of simulated
data point in each histogram bin 
}
}
\examples{
###################
## Example 1     ##
###################
x <- seq(-3, 3, length.out=100) ## Data
samp <- rnorm(1e6)              ## Monte Carlo simulation 
h <- 0.8*bw.nrd0(samp)          ## Define bandwidth using R's bw.nrd0

## First, I demonstrate how to use point-wise logLik
## Note logLik_pw returns log-likelihood.
system.time(pw1  <- logLik_pw(x, samp, h=h, m=1))
##   user  system elapsed 
##  3.480   0.120   3.605 

## Second, I demonstrate how to use KDE-FFT. logLik_fft2 retrieves 
## log-likelihoods for all data point  
system.time(fft1 <- logLik_fft2(x, samp, h=h, m=1)[["PDF"]])
##   user  system elapsed 
##  1.056   0.024   1.079 

## Third, I demonstrate how to build-in routine binding logLik_fft and 
## rnorm to get Gaussian density. Enter ?logLik_norm2 to see help page
## pVec stands for parameter vector. For a Gaussian model, pVec stores mean
## and standard deviation.
system.time(fft2 <- logLik_norm2(x, pVec=c(0,1), 1e6, h=h, m=1)[["PDF"]])
##   user  system elapsed 
##  1.304   0.040   1.349  

## You should get all 4 lines overlaping one another
## the tails may be a bit off.
plot(x, pw1,type="l", lty="dotted")
lines(x, fft1, col="darkgreen", lty="dashed")
lines(x, fft2, col="blue", lty="dotdash")
lines(x, dnorm(x, log=TRUE), col="red")

###################
## Example 2     ##
###################
## Use piecewise LBA data as an example
data(lba)
logLik_fft(plba$DT1, plba$eDT1)
logLik_fft(plba$DT2, plba$eDT2)
plbaEg1 <- logLik_fft2(plba$DT1, plba$eDT1)
plbaEg2 <- logLik_fft2(plba$DT2, plba$eDT2)

str(plbaEg1)
## List of 4
## $ LL      : num 280
## $ PDF     : num [1:695, 1] 0.918 -0.456 0.668 0.872 0.788 ...
## $ z       : num [1:1024, 1] 0.128 0.129 0.13 0.131 0.133 ...
## $ PDF_hist: num [1:1024, 1] 0 0 0 0 0 0 0 0 0 0 ...
str(plbaEg2)
## List of 4
## $ LL      : num 45.3
## $ PDF     : num [1:305, 1] 0.6396 -0.1174 0.6178 -1.3723 -0.0273 ...
## $ z       : num [1:1024, 1] 0.121 0.124 0.126 0.129 0.131 ...
## $ PDF_hist: num [1:1024, 1] 0 0 0 0 0 0 0 0 0 0 ...
}
\references{
Holmes, W. (2015). A practical guide to the Probability Density
Approximation (PDA) with improved implementation and error characterization.
\emph{Journal of Mathematical Psychology}, \bold{68-69}, 13--24,
doi: http://dx.doi.org/10.1016/j.jmp.2015.08.006.
}
\seealso{
\code{\link{logLik_pw}}, \code{\link{logLik_fft}},
\code{\link{bw.nrd}}, \code{\link{bandwidth.nrd}}.
}

