% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_lba}
\alias{logLik_lba}
\title{Compute FFT Log-likelihood for a LBA Model}
\usage{
logLik_lba(object, pVec, ns)
}
\arguments{
\item{object}{a matrix storing empirical choice RT data. First column must
stores choices and second column stores RTs in seconds.}

\item{pVec}{pLBA parameter vector. The sequence is 
critical. It is, b1, b2, A1, A2, mu1, mu2, sigma1, sigma2, t01, and t02.}

\item{ns}{number of simulations. Usually \code{cpda} can handle up to 1e6. 
Use \code{gpda} if large simulation is required.}
}
\value{
a summed, logged likelihood across trials and accumulators.
}
\description{
This is a wrapper function to approximate a LBA distribution, using
KDE-FFT method. To retrieve more outputs, please use \code{logLik_LBA2}.
}
\examples{
 
rm(list=ls())
data(lba)
dMat <- data.matrix(d)
head(dMat)
 
pVec <- c(b1=1, b2=1, A1=.5, A2=.5, mu1=2.4, mu2=1.6, sigma1=1, sigma2=1.2,
t01=.5, t02=.5)
tmp0 <- cpda::logLik_lba(dMat, pVec, 1e5); tmp0
 
}

