% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_norm}
\alias{logLik_norm}
\alias{logLik_norm2}
\title{Compute FFT Log-likelihood for a Gaussian Distribution}
\usage{
logLik_norm(object, pVec, ns)

logLik_norm2(object, pVec, ns)
}
\arguments{
\item{object}{a vector storing empirical data}

\item{pVec}{parameter vector storing mean and standard deviation}

\item{ns}{number of simulations.}
}
\value{
Log-likelihood; plus PDF, grid centers, and PDF_hist
}
\description{
This is a wrapper function to approximate a Gaussian distribution, using
KDE-FFT method. To retrieve more outputs and replicate Holmes's example 1,
please use \code{logLik_norm2}.
}
\examples{
pVec <- c(mu=5, sigma=1)
y    <- sort(rnorm(1e5, pVec[1], pVec[2]))

ll1 <- logLik_norm(y, pVec, 1e5)
ll2 <- logLik_norm2(y, pVec, 1e5)
str(ll1); str(ll2)

plot(ll2$z, ll2$PDF_hist, type="l", lty=2,
main="Normal Distribution",xlab="x",ylab="L(x|beta)")
lines(y, ll2$PDF, col="red", lwd = 1.5)

}

