% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_pLBA}
\alias{logLik_pLBA}
\title{Compute FFT Log-likelihood for Piece-wise LBA Model}
\usage{
logLik_pLBA(data, pVec, MCMC_params)
}
\arguments{
\item{data}{subject data}

\item{pVec}{a parameter vector for piece-wise LBA Model}

\item{MCMC_params}{MCMC parameters}
}
\value{
Log-likelihood
}
\description{
Use logLik_fft to get probability density for a piecewise LBA model.
}
\examples{
mcmcParams <- list(sigma_exact=1, bandwidth=.02, LL_NSAMPLE=1e4,
                  Nstep=30, Nchain=24, noise_size=.001, burnin=10,
                  resample_mod=3)

load("data/Data1.rda")
dplyr::tbl_dt(d)

## Source: local data table [1,000 x 3]
##      Response ResponseTime Block
##         (dbl)        (dbl) (dbl)
##   1         0    0.5676343     1
##   2         1    0.6183177     1
##   3         0    0.8806298     1
##   4         0    0.4563023     1
##   5         1    0.8496136     1
##   6         0    0.5866219     1
##   7         1    0.6482302     1
##   8         1    0.3510035     1
##   9         0    0.6117150     1
##   10        0    0.6940521     1
##   ..      ...          ...   ...

pVec <- c(1.51, 3.32, 1.51, 2.24, 3.69, 0.31, 0.08)
tmp0 <- Compute_log_likelihood_FFT(d, pVec, mcmcParams)
tmp1 <- logLik_pLBA(d, pVec, mcmcParams)

require(rbenchmark)
within(benchmark(R=Compute_log_likelihood_FFT(d, pVec, mcmcParams),
  Cpp=logLik_pLBA(d, pVec, mcmcParams),
  replications=rep(1e2, 3),
  columns=c('test', 'replications', 'elapsed'),
  order=c('test', 'replications')),
  { average = elapsed/replications })
##     test replications elapsed average
##  2  Cpp          100   0.736 0.00736
##  4  Cpp          100   0.737 0.00737
##  6  Cpp          100   0.736 0.00736
##  1    R          100   1.037 0.01037
##  3    R          100   0.922 0.00922
##  5    R          100   0.922 0.00922

}

