% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_prior}
\alias{logLik_prior}
\title{Prior Log-likelihood}
\usage{
logLik_prior(pVec)
}
\arguments{
\item{pVec}{a vector storing pLBA model parameters. The sequence is 
critical. It is, A1, A2, v1, v2, t01, t02.}
}
\value{
a prior log prior likelihood
}
\description{
This function computes the prior log-likelihood for the given parameter set,
using uniform distributions.
}
\examples{
logLik_prior(pVec)

}
