% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pda.R
\docType{package}
\name{pda}
\alias{pda}
\alias{pda-package}
\title{Probability Density Approximation}
\description{
Provide C++ enhanced computation for probability density approximation
}
\author{
Yi-Shin Lin <yishin.lin@utas.edu.au> \cr
Andrew Heathcote <andrew.heathcote@utas.edu.au>
}
\keyword{pda}

