% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pick_2chains}
\alias{pick_2chains}
\title{Pick Two Other Chains Randomly}
\usage{
pick_2chains(k, chains)
}
\arguments{
\item{k}{an integer indicating which chain is currently running. This has
to be an integer and with the range of 0 to nchain-1. There is no check to
preventing the function from crashing, if the user enter an irregular
number.}

\item{chains}{an integer vector (std::vector) from 0 to nchain-1.}
}
\value{
an Armadillo vector
}
\description{
This is part of the algorithm of DE-MC. The function picks two other
chains, except the currently processed one.
}
\examples{
nchain <- 24
chainSeq <- (1:24)-1
pick_2chains(3, chainSeq)
}
\keyword{pick_2chains}

