% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pickchains}
\alias{pickchains}
\title{Pick Other Chains Randomly}
\usage{
pickchains(k, n, chains)
}
\arguments{
\item{k}{an integer indicating which chain is processed currently. This 
must be an integer within the range of 0 to \code{nchain-1} (i.e., C index).
No check for errorly using R index, because this is an internal function.}

\item{n}{the numbers of chain to sample.}

\item{chains}{a numeric vector of chain index, e.g., 0:24.}
}
\value{
a column vector
}
\description{
This is part of DEMC algorithm. The function randomly samples \code{n}   
chains out of \code{length(chains)} chains, excluding the kth chain.
chains.
}
\examples{
nchain <- 24
chainSeq <- (1:24)-1 ## Convert to C index

## Current processing chain is the fourth chain (index=3) 
## We wish to pick 2 chains out of 24 chains
pickchains(3, 2, chainSeq) 

}
\keyword{pickchains}
