% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rlba}
\alias{rlba}
\title{Generate Random Choice Response Times Using LBA Model}
\usage{
rlba(n, pVec)
}
\arguments{
\item{n}{number of observations. Must be an integer}

\item{pVec}{a two-accumulator LBA parameter vector. The sequence is 
critical. It is, b1, b2, A1, A2, mu1, mu2, sigma1, sigma2, t01, and t02.}
}
\value{
A matrix. First column stores choices and second column stores RTs 
in seconds
}
\description{
This function uses two-accumulator LBA model to generate random choice RTs.
}
\details{
\itemize{
\item \bold{\emph{b1}} response threshold for accumulator 1. 
\item \bold{\emph{b2}} response threshold for accumulator 2. 
\item \bold{\emph{A1}} starting point interval for accumulator 1. 
\item \bold{\emph{A2}} starting point interval for accumulator 2. 
\item \bold{\emph{mu1}} drift rate for accumulator 1.
\item \bold{\emph{mu2}} drift rate for accumulator 2. 
\item \bold{\emph{sigma1}} drift rate standard deviation for accumulator 1.
\item \bold{\emph{sigma2}} drift rate standard deviation for accumulator 2.
\item \bold{\emph{t01}} non-decision time for accumulator 1. 
\item \bold{\emph{t02}} non-decision time for accumulator 2. 
}
}
\examples{
n    <- 10
pVec <- c(b1=1, b2=1, A1=.5, A2=.5, mu1=2.4, mu2=1.6, sigma1=1, sigma2=1.2,
          t01=.5, t02=.5)
dat1 <- cpda::rlba(n, pVec)
str(dat1)
summary(dat1)

## Compare to rtdists
## dat2 <- rtdists::rLBA(n, A=0.5, b=1, t0 = 0.5, mean_v=c(2.4, 1.6), 
## sd_v=c(1,1.2))
## str(dat2)
## summary(dat2)

}

