% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rplba}
\alias{rplba}
\title{Generate Random Choice Decision Times}
\usage{
rplba(n, pVec)
}
\arguments{
\item{n}{number of observations. Must be an integer}

\item{pVec}{a double vector storing pLBA model parameters. The sequence is 
critical. It is, A, b, muv1, muv2, t_ND, muw1, muw2, t_delay, sv, swt.}
}
\value{
A two-element list with estimated decision time for accumulator 1,
accumultor 2, empirical decision time for accumualtor 1 and accumulator 2.
}
\description{
Uses two-accumulator pLBA model to generate random choice DTs (not RTs!).
}
\details{
\itemize{
\item \bold{\emph{A}} starting point interval. A starting point is with the 
interval \code{[0, A]}. Average amount of prior evidence (ie before accumulation 
process even begins) across trials is \code{A/2}. 
\item \bold{\emph{b}} response threshold. \code{b-A/2} is a measure of 
\emph{response caution}.
\item \bold{\emph{muv1}} accumulator 1 drift rate, piece 1
\item \bold{\emph{muv2}} accumulator 2 drift rate, piece 1
\item \bold{\emph{t_ND}} non-decision time in second. 
\item \bold{\emph{muw1}} accumulator 1 drift rate, piece 2
\item \bold{\emph{muw2}} accumulator 2 drift rate, piece 2
\item \bold{\emph{t_delay}} delay time for the two-piece process   
\item \bold{\emph{sv}} a common standard deviation for all drift rates 
(muv1, muw1, muv2, muw2).
\item \bold{\emph{swt}} switch time, usually determined by experimental 
design).
}
}
\examples{
pVec <- c(A=1.51, b=2.7, muv1=3.32, muv2=2.24, t_ND=0.08, muw1=1.51, 
          muw2=3.69, t_delay=0.31, sv=1, swt=0.5)
## setting <- c(bandwidth=.02, ns=1e5, nmc=30, nchain=24, rp=.001, burnin=10,
## nthin=3, start=1, gammaMult=2.38, report=100)
DTs <- cpda::rplba(n=1e4, pVec)
summary(DTs)
head(DTs)

}

