% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run}
\alias{run}
\title{Use DEMC Sampler to Fit a Hierarchical Bayesian Model}
\usage{
run(MCMC_params, Model_specifics, report = 100L)
}
\arguments{
\item{MCMC_params}{Provides all MCMC and KDE parameters along with a
few other things}

\item{Model_specifics}{Stores function handles and block information
for the model}
}
\value{
0
}
\description{
Run DEMC-EAM Fit
}
\examples{
load("data/Data1.rda")
dplyr::tbl_dt(d)
mcmcParams <- list(sigma_exact=1, bandwidth=.02, LL_NSAMPLE=1e4,
Nstep=30, Nchain=24, noise_size=.001, burnin=10, resample_mod=3,
start=1)
## Group parameters into MCMC blocks; should follow the order
mod <- list(block_1_ind=c(1,2,4,7),  ## A, muv1, muv2, t_er
             block_2_ind=c(3,5,6))    ## muw1, muw2, t_delay
tmp0 <- run(MCMC_params=mcmcParams, Model_specifics=mod)
}
\keyword{demc}

