% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pLBA.R
\name{SwitchModel_Prior}
\alias{SwitchModel_Prior}
\title{Switch Model's Prior Distributions}
\usage{
SwitchModel_Prior(pVec)
}
\arguments{
\item{pVec}{a parameter vector. Similar to DMC's p.vector}
}
\value{
a prior probability density summed across parameters
}
\description{
This function computes the prior for the given parameter set. The prior
for each parameter here is just a uniform distribution for simplicity.
}
\examples{
p.vector <- c(1.51, 3.32, 1.51, 2.24, 3.69, 0.31, 0.08)
SwitchModel_Prior(pVec=p.vector)
}

