% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SwitchModelPrior}
\alias{SwitchModelPrior}
\title{Switch Model's Prior Distributions}
\usage{
SwitchModelPrior(pVec)
}
\arguments{
\item{pVec}{a parameter vector. Similar to DMC's p.vector}
}
\value{
a prior probability density summed across parameters
}
\description{
This Rcpp function computes the prior for the given parameter set. The
prior for each parameter here is just a uniform distribution for
simplicity.
}
\examples{
p.vector <- c(1.51, 3.32, 1.51, 2.24, 3.69, 0.31, 0.08)
SwitchModelPrior(pVec=p.vector)
require(rbenchmark)
within(benchmark(R=SwitchModel_Prior(p.vector),
                 Cpp=SwitchModelPrior(p.vector),
                 replications=rep(1e4, 3),
                 columns=c('test', 'replications', 'elapsed'),
                 order=c('test', 'replications')),
                 { average = elapsed/replications })
#   test replications elapsed  average
# 2  Cpp        10000   0.048 4.80e-06
# 4  Cpp        10000   0.032 3.20e-06
# 6  Cpp        10000   0.056 5.60e-06
# 1    R        10000   0.151 1.51e-05
# 3    R        10000   0.182 1.82e-05
# 5    R        10000   0.142 1.42e-05
}

