% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossover}
\alias{crossover}
\alias{getsubchains}
\alias{migration}
\title{Crossover and Migration Samplers}
\usage{
crossover(theta, gammamult = 2.38, rp = 0.001)

getsubchains(nchain)

migration(theta, gammamult = 2.38, rp = 0.001)
}
\arguments{
\item{theta}{a npar x nchain matrix}

\item{rp}{ter Braak's (2006) b, setting the range of the uniform 
distribution that derives epsilon. This is set at 1e-3.}

\item{gamma}{a gamma vector.}

\item{k}{current chain index. Note chain index starts from 0.}
}
\value{
a column vector
}
\description{
This is part of DE-MC algorithm (crossover) and Distributed Evolutionary 
Monte Carlo (migration). Both samplers propose a set of new parameters.
}
\examples{
init <- function(nc, npar) {
  theta0  <- array(dim = c(nc, npar))  
  theta0[,2] <- runif(nc, 0, 10);   # b
  theta0[,1] <- runif(nc, 0, theta0[,2]); # A
  theta0[,3] <- runif(nc, 0, 7);   # v1
  theta0[,4] <- runif(nc, 0, 7);   # v2
  theta0[,5] <- runif(nc, 0, .5)   # t0           
  return(theta0)
}

npar <- 5 
nc   <- npar * 1 
theta0  <- init(nc, npar)
gamma <- cpda::gammavec(npar, 2.38)
  
## Current chain index is 0 (C-based index)
crossover(theta0, gamma, 0)
crossover(theta0, gamma, 1)
crossover(theta0, gamma, 2)
crossover(theta0, gamma, 3)
crossover(theta0, gamma, 4)

migration(theta0, gamma, 4)
get_subchains(nc)

}
\references{
ter Braak, C. J. F. (2006). A Markov Chain Monte Carlo version 
of the genetic algorithm Differerntial Evolution: Easy Bayesian computing
for real parameter spaces. Statistics and Computing, 16, 239–249. \cr\cr
Hu, B., & Tsui, K.-W. (2005). Distributed evolutionary Monte Carlo with 
applications to Bayesian analysis Technical Report Number 1112.
}
