% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm.R
\name{dtnorm}
\alias{dtnorm}
\alias{rtnorm}
\title{Truncated Normal Distribution}
\usage{
dtnorm(x = 0, mean = 0, sd = 1, lower = -Inf, upper = Inf, log = 0,
  checks = FALSE)

rtnorm(n = 1, mean = 0, sd = 1, lower = -Inf, upper = Inf,
  checks = FALSE)
}
\arguments{
\item{x, n}{x in \code{dtnorm} is a vector of quantiles; n in
\code{rtnorm} indicates how many random numbers to generate}

\item{mean}{a scalar or a vector of means}

\item{sd}{a scalar or a vector of standard deviations}

\item{lower}{lower truncation value Default as -Inf (ie regular Gaussian)}

\item{upper}{upper truncation value. Default as Inf (ie regular Gaussian).}

\item{log}{default 0 (FALSE). Enter 1 to make it TRUE. This will return
logged likelihood. This argumement is used only in \code{dtnorm}.}

\item{checks}{a switch to turn on check functions to check inputs and ouput.
Default to FALSE.}
}
\description{
The two wrapper functions call two corresponding Rcpp functions to speed up
computation. The codes are based on Christopher Jackson's \pkg{msm}
(1.5) package and Jonathan Olmsted's \pkg{RcppTN} (0.1-8) package.
}
\details{
\code{dtnorm} calculates probability density for a truncated normal
distribution with mean equal to \code{mean} and standard deviation equal to
\code{sd} and truncates at \code{lower} and \code{upper} range.

\code{rtnorm} generates a random number from a truncated Normal
distribution with \code{mean} and \code{sd}.
}
\examples{
## Use dtnorm and rtnorm with their default values
dtnorm()
rtnorm()

## A similar curve plotting example extracted from dnorm functions
plot(function(x) dnorm(x, log = FALSE), -2.5, 2.5,
     main = "Normal Distribution", ylim=c(0,0.45), ylab="Density")
curve(dtnorm(x, lower=-2, upper=2), add=TRUE, col="tomato", lwd=2)
mtext("dnorm(x)", adj = 0)
mtext("dtnorm(x)", col = "tomato", adj = 1)
}

