% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gammaVec}
\alias{gammaVec}
\title{Generate a Gamma Vector}
\usage{
gammaVec(n, gamma)
}
\arguments{
\item{n}{number of parameters.}

\item{gamma}{a tuning parameter for gamma mutation. Usually set at 2.38}
}
\value{
a column vector
}
\description{
This is part of DEMC algorithm. \code{gammaVec} generates a gamma vector to 
facilitate vector-based element-wise compuation at C++.
}
\examples{
pVec <- c(A=1.51, b=2.7, muv1=3.32, muv2=2.24, t_ND=0.08, muw1=1.51,  
          muw2=3.69, t_delay=0.31,  sv=1, swt=0.5)
gamma <- gammaVec(length(pVec), 2.38)
}

