% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{init}
\alias{init}
\title{Initialise a Piecewise LBA Sample}
\usage{
init(data, MCMC_params, Model_specifics, gammaMult = 2.38)
}
\arguments{
\item{data}{the choice-RT data from subjects to be fit.}

\item{MCMC_params}{Provides all MCMC and KDE parameters along with a
few other things}

\item{Model_specifics}{Stores function handles and block information
for the model}

\item{gammaMult}{default 2.38}

\item{report}{the iteration interval of returning a progress report.}
}
\value{
the chain parameter information (param_chain). This ends up
saved in a data file however. The matrix has a structure
\code{param_chain(MCMC_iteration,parameter num,chain num)}
Default 100
}
\description{
Generate a Bayesian initial sample
}
\examples{
load("data/Data1.rda")
dplyr::tbl_dt(d)
mcmcParams <- list(sigma_exact=1, bandwidth=.02, LL_NSAMPLE=1e4,
Nstep=30, Nchain=24, noise_size=.001, burnin=10, resample_mod=3,
start=1)
## Group parameters into MCMC blocks; should follow the order
mod <- list(block_1_ind=c(1,2,4,7),  ## A, muv1, muv2, t_er
             block_2_ind=c(3,5,6))    ## muw1, muw2, t_delay
tmp0 <- init(data=d, MCMC_params = mcmcParams, Model_specifics = mod)
}
\keyword{demc}

