% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_fft}
\alias{logLik_fft}
\title{Compute Log-likelihood Using Fast Fourier Transform}
\usage{
logLik_fft(DT, eDT, m, M, h, ns)
}
\arguments{
\item{DT}{a vector of empirical decision times}

\item{eDT}{a vector of modelled decision times}

\item{m}{min value}

\item{M}{max value}

\item{h}{KDE bandwidth}

\item{ns}{number of binned samples}
}
\value{
Log-likelihood
}
\description{
This function implements Holmes's (2015) KDE-FFT method to calculate
approximate probability density.
}
\examples{
DT1  <- read.csv("data/DT1.csv", header=F)
DT2  <- read.csv("data/DT2.csv", header=F)
eDT1 <- read.csv("data/eDT1.csv", header=F)
eDT2 <- read.csv("data/eDT2.csv", header=F)
bandwidth <- .02;
nsample <- 1e4
m <- min(c(DT1$V1, DT2$V1)) - 3 * bandwidth
M <- max(c(DT1$V1, DT2$V1)) + 3 * bandwidth
logLik_fft(DT1$V1, eDT1$V1, m, M, bandwidth, nsample)
logLik_fft(DT2$V1, eDT2$V1, m, M, bandwidth, nsample)
}
\references{
Holmes, W. (2015). A practical guide to the Probability Density
Approximation (PDA) with improved implementation and error characterization.
\emph{Journal of Mathematical Psychology}, \bold{68-69}, 13--24,
doi: http://dx.doi.org/10.1016/j.jmp.2015.08.006.
}

