% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLik_plba}
\alias{logLik_plba}
\title{Compute FFT Log-likelihood for a pLBA Model}
\usage{
logLik_plba(object, pVec, ns)
}
\arguments{
\item{object}{a matrix storing empirical choice RT data. First column must
stores choices and second column stores RTs in seconds.}

\item{pVec}{a vector storing pLBA model parameters. The sequence is 
critical. It is, A, b, muv1, muv2, t_ND, muw1, muw2, t_delay, sv, swt.}

\item{ns}{number of simulations. Usually \code{cpda} can handle up to 1e6. 
Use \code{gpda} if large simulation is required.}
}
\value{
a summed, logged likelihood across trials and accumulators.
}
\description{
This is a wrapper function to approximate a pLBA distribution, using
KDE-FFT method. To retrieve more outputs, please use \code{logLik_pLBA2}.
}
\examples{
 
rm(list=ls())
data(lba)
dMat <- data.matrix(d)
head(dMat)

pVec <- c(A=1.51, b=2.7, muv1=3.32, muv2=2.24, t_ND=0.08, muw1=1.51, muw2=3.69,
          t_delay=0.31, sv=1, swt=0.5)
## setting <- c(bandwidth=.02, ns=1e5, nmc=30, nchain=24, rp=.001, burnin=10,
## nthin=3, start=1, gammaMult=2.38, report=100)
tmp0 <- cpda::logLik_plba(dMat, pVec, 1e5); tmp0
 
}

