% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pickchains}
\alias{pickchains}
\title{Pick Two Other Chains Randomly}
\usage{
pickchains(k, n, chains)
}
\arguments{
\item{k}{an integer indicating which chain is currently running. This has
to be an integer and with the range of 0 to nchain-1. There is no check to
preventing the function from crashing, if the user enter an irregular
number.}

\item{n}{number of picked chains.}

\item{chains}{a vector of chain index, e.g., 0:24.}
}
\value{
a column vector
}
\description{
This is part of DEMC algorithm. The function picks three other chains (0, 
1, 2), except the currently processed one (k).
}
\examples{
nchain <- 24
chainSeq <- (1:24)-1 ## Use C++ index
pickchains(3, 2, chainSeq)

}
\keyword{pickchains}

