% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rplba}
\alias{rplba}
\title{Retrieve Empirical and Modelled Decision Times}
\usage{
rplba(data, pVec, setting)
}
\arguments{
\item{data}{a Rcpp List for subject data}

\item{pVec}{a Rcpp NumvericVector for parameter vector}

\item{setting}{a named numeric vector storing DE and MCMC parameters:
sigma_exact, bandwidth, ns (nsample), nmc, nchain, rp (DE noise,
usually .001), burnin, nthin, start (usually 1), gammaMult (DE gamma,
usually 2.38), ST (piecewise LBA switch time, depending on experimental
design), and report}
}
\value{
A four-element list with estimated decision time for accumulator 1,
accumultor 2, empirical decision time for accumualtor 1 and accumulator 2.
}
\description{
Get decision times for two-accumulator LBA model
}
\examples{
data(lba)
dMat <- data.matrix(d)

##  A       muv1    muw1    muv2    muw2    t_delay t_ND    b
p.vector <- c(A=1.51, muv1=3.32, muw1=1.51, muv2=2.24, muw2=3.69,
              t_delay=0.31, t_ND=0.08, b=2.7)
setting <- c(sigma_exact=1, bandwidth=.02, ns=1e2, nmc=30, nchain=24,
             rp=.001, burnin=10, nthin=3, start=1, gammaMult=2.38,
             ST=attr(d, "SwitchTime"), report=100)

DTs <- pda::rplba(dMat, p.vector, setting)
str(DTs)

## List of 4
## $ eDT1: num [1:72, 1] 0.317 0.425 0.71 0.345 0.526 ...
## $ eDT2: num [1:28, 1] 0.45 0.418 0.56 0.719 0.313 ...
## $ DT1 : num [1:695, 1] 0.488 0.801 0.376 0.507 0.532 ...
## $ DT2 : num [1:305, 1] 0.538 0.77 0.568 0.271 0.881 ...
}

